# Improve performance by caching conversion factors