--- a/UnitConversionBenchmark/Program.cs
+++ b/UnitConversionBenchmark/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using UnitConversionBenchmark;
+
+var summary = BenchmarkRunner.Run<CompareUnitConversionBenchmark>();