--- a/UnitConversionBenchmark/UnitConversionBenchmark.csproj
+++ b/UnitConversionBenchmark/UnitConversionBenchmark.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+<PropertyGroup>
+  <OutputType>Exe</OutputType>
+  <TargetFramework>net6.0</TargetFramework>
+  <ImplicitUsings>enable</ImplicitUsings>
+  <Nullable>enable</Nullable>
+</PropertyGroup>
+
+<ItemGroup>
+  <AdditionalFiles Include="..\stylecop.json" Link="stylecop.json" />
+</ItemGroup>
+
+<ItemGroup>
+  <ProjectReference Include="..\UnitConversion\UnitConversion.csproj" />
+</ItemGroup>
+
+<ItemGroup>
+  <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
+</ItemGroup>
+
+</Project>