… UnitConversionExample/Properties/AssemblyInfo.cs
… UnitConversionTests/Properties/AssemblyInfo.cs