Improve performance for multiple conversations by caching conversion factors. Added a Benchmark project to proof it.

**Before**
|                    Method | Count |           Mean |         Error |        StdDev |
|-------------------------- |------ |---------------:|--------------:|--------------:|
| ConvertWithUnitConversion |    10 |     6,998.9 ns |     138.77 ns |     228.01 ns |
|   ConvertWithDirectFactor |    10 |       226.4 ns |       3.91 ns |       3.66 ns |
| ConvertWithUnitConversion |   100 |    68,063.3 ns |   1,139.99 ns |   1,170.69 ns |
|   ConvertWithDirectFactor |   100 |     1,352.7 ns |      24.93 ns |      23.32 ns |
| ConvertWithUnitConversion |  1000 |   662,034.8 ns |  12,972.85 ns |  13,322.16 ns |
|   ConvertWithDirectFactor |  1000 |    12,603.1 ns |     243.87 ns |     271.06 ns |
| ConvertWithUnitConversion | 10000 | 6,873,128.6 ns | 132,366.27 ns | 189,835.75 ns |
|   ConvertWithDirectFactor | 10000 |   124,244.3 ns |   1,618.71 ns |   1,514.14 ns |

**After**
|                    Method | Count |         Mean |     Error |    StdDev |
|-------------------------- |------ |-------------:|----------:|----------:|
| ConvertWithUnitConversion |    10 |     245.9 ns |   2.37 ns |   2.22 ns |
|   ConvertWithDirectFactor |    10 |     233.3 ns |   1.24 ns |   1.16 ns |
| ConvertWithUnitConversion |   100 |   1,598.3 ns |  11.68 ns |  10.93 ns |
|   ConvertWithDirectFactor |   100 |   1,383.4 ns |   8.22 ns |   7.69 ns |
| ConvertWithUnitConversion |  1000 |  14,749.3 ns |  24.61 ns |  21.82 ns |
|   ConvertWithDirectFactor |  1000 |  12,812.5 ns |  65.93 ns |  61.67 ns |
| ConvertWithUnitConversion | 10000 | 145,785.3 ns | 550.44 ns | 514.88 ns |
|   ConvertWithDirectFactor | 10000 | 125,695.9 ns | 397.88 ns | 352.71 ns |